tls: split ServerContextConfigImpl into it's own compilation unit
