+++ source/extensions/transport_sockets/tls/BUILD
@@ -31,6 +31,7 @@ envoy_cc_library(
         "//envoy/network:transport_socket_interface",
         "//envoy/registry",
         "//envoy/server:transport_socket_config_interface",
+        "//source/common/tls:server_context_config_lib",
         "//source/common/tls:server_ssl_socket_lib",
         "@envoy_api//envoy/extensions/transport_sockets/tls/v3:pkg_cc_proto",
     ],
@@ -47,6 +48,7 @@ envoy_cc_library(
         "//envoy/registry",
         "//envoy/server:transport_socket_config_interface",
         "//source/common/tls:client_ssl_socket_lib",
+        "//source/common/tls:context_config_lib",
         "@envoy_api//envoy/extensions/transport_sockets/tls/v3:pkg_cc_proto",
     ],
     alwayslink = True,
