+++ source/extensions/transport_sockets/tls/downstream_config.cc
@@ -4,7 +4,7 @@
 #include "envoy/extensions/transport_sockets/tls/v3/tls.pb.validate.h"
 
 #include "source/common/protobuf/utility.h"
-#include "source/common/tls/context_config_impl.h"
+#include "source/common/tls/server_context_config_impl.h"
 #include "source/common/tls/server_ssl_socket.h"
 
 namespace Envoy {
