+++ test/common/tls/BUILD
@@ -47,6 +47,7 @@ envoy_cc_test(
         "//source/common/stream_info:stream_info_lib",
         "//source/common/tls:context_config_lib",
         "//source/common/tls:context_lib",
+        "//source/common/tls:server_context_config_lib",
         "//source/common/tls:server_context_lib",
         "//source/common/tls:ssl_socket_lib",
         "//source/common/tls:utility_lib",
@@ -92,6 +93,7 @@ envoy_cc_test(
         "//source/common/stats:stats_lib",
         "//source/common/tls:context_config_lib",
         "//source/common/tls:context_lib",
+        "//source/common/tls:server_context_config_lib",
         "//source/common/tls:server_context_lib",
         "//test/common/tls/test_data:cert_infos",
         "//test/mocks/init:init_mocks",
@@ -196,8 +198,8 @@ envoy_cc_test(
     srcs = ["handshaker_factory_test.cc"],
     deps = [
         "//source/common/stream_info:stream_info_lib",
-        "//source/common/tls:context_config_lib",
         "//source/common/tls:context_lib",
+        "//source/common/tls:server_context_config_lib",
         "//source/common/tls:server_context_lib",
         "//source/common/tls:ssl_handshaker_lib",
         "//source/common/tls:ssl_socket_lib",
