+++ test/common/tls/handshaker_factory_test.cc
@@ -6,8 +6,8 @@
 
 #include "source/common/stream_info/stream_info_impl.h"
 #include "source/common/tls/client_ssl_socket.h"
-#include "source/common/tls/context_config_impl.h"
 #include "source/common/tls/context_manager_impl.h"
+#include "source/common/tls/server_context_config_impl.h"
 #include "source/common/tls/ssl_handshaker.h"
 #include "source/server/process_context_impl.h"
