+++ test/integration/alpn_selection_integration_test.cc
@@ -3,8 +3,8 @@
 #include "envoy/extensions/transport_sockets/tls/v3/cert.pb.h"
 
 #include "source/common/http/utility.h"
-#include "source/common/tls/context_config_impl.h"
 #include "source/common/tls/context_impl.h"
+#include "source/common/tls/server_context_config_impl.h"
 #include "source/common/tls/server_ssl_socket.h"
 
 #include "test/integration/http_integration.h"
