+++ test/integration/ssl_utility.cc
@@ -6,8 +6,8 @@
 #include "source/common/json/json_loader.h"
 #include "source/common/network/utility.h"
 #include "source/common/tls/client_ssl_socket.h"
-#include "source/common/tls/context_config_impl.h"
 #include "source/common/tls/context_manager_impl.h"
+#include "source/common/tls/server_context_config_impl.h"
 #include "source/common/tls/server_ssl_socket.h"
 
 #include "test/config/utility.h"
